Renamed the SgtUrlBreadCrumb to SgtBreadCrumb
